+++ code/java/Pattern8.java
@@ -1,5 +1,5 @@
 
-public class patt8 {
+class patt8 {
     public static void main(String[] args) {
         int r =5;
         for (int i = 1; i <= r; i++) {
